statement_timeout – no opinionated recommendation
Saying that global statement_timeout is not recommended is too opinionaned – in OLTP systems, it's not uncommon to have limit for HTTP processing time set to 30 or 60 seconds. If the primary goal of DBMS is supporting HTTP request processing (websites, mobile apps, APIs), then having statement_timeout set to 60s, 30s, or even less is normal, protecting from long-running and orphaned queries, focusing on very fast processing. Documentation should be less opinionated not to scary those who consider this option.