Statistics on dirtied and allocated local buffers.
It is a preliminary commit that tracks the state of temp buffers. The main goal
of this statistic is to provide the optimiser with numbers to compute
the cost estimation of flushing temporary buffers.

Such a flush may be necessary if the optimiser decides to build a plan that
includes a parallel section of the query, which involves scanning
a temporary table.